Use float DF rope callback phases

+++ include/main/dll/DF/dll_194.h
@@ -5,8 +5,8 @@
 
 int dfropenode_func0E(int obj, f32 x, f32 y, f32 z, float *distanceOut, float *phaseOut,
                       u8 *sideOut);
-void dfropenode_render2(double phase, double force, int obj);
-void dfropenode_modelMtxFn(double distance, int obj, float *phase);
-void dfropenode_func0B(double phase, int obj, float *xOut, float *yOut, float *zOut);
+void dfropenode_render2(f32 phase, f32 force, int obj);
+void dfropenode_modelMtxFn(f32 distance, int obj, float *phase);
+void dfropenode_func0B(f32 phase, int obj, float *xOut, float *yOut, float *zOut);
 
 #endif /* MAIN_DLL_DF_DLL_194_H_ */
